1. id in blob ( labelBlob ) is changed to int32_t 2. cell is id is in the order at which they are created. 3. cell id is copied in to all of its contractions 4. guy uses the above to create a tree node of cell/contraction information

+++ svl/src/labelBlob.cpp
@@ -216,7 +216,7 @@ svl::labelBlob::labelBlob(const cv::Mat& gray, const cv::Mat& threshold_out, con
     reload(gray, threshold_out, client_id,  minAreaPixelCount);
 }
 
-svl::labelBlob::ref labelBlob::create(const cv::Mat& gray, const cv::Mat& threshold_out, const int64_t client_id, const int minAreaPixelCount) {
+svl::labelBlob::ref labelBlob::create(const cv::Mat& gray, const cv::Mat& threshold_out, const int32_t client_id, const int minAreaPixelCount) {
     auto reff = std::make_shared<labelBlob> (gray, threshold_out, minAreaPixelCount, client_id);
     return reff;
 }
